+++ job_offers_data_load/PostgreSQL_data_insert.py
@@ -3,7 +3,7 @@ import sqlalchemy.orm
 import sqlalchemy.sql.schema
 from sqlalchemy import select
 import datetime
-from JobOffers_DataLoad.PostgreSQL_tables_declaration import *
+from job_offers_data_load.PostgreSQL_tables_declaration import *
 
 
 def is_already_for_list_like(dataframe: pandas.DataFrame,
