+++ job_offers_data_load/PostgreSQL_tables_declaration.py
